requirements.txt
Flask
python-dotenv
azure-core
azure-ai-formrecognizer
azure-storage-blob
azure-search-documents
google-cloud-storage
langchain
